--- Disuku.MongoStorage/IMongoStorage.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Disuku.MongoStorage
-{
-    public interface IMongoDbStorage
-    {
-        void InitializeDataBase(string database);
-        bool Exists();
-        void Store<T>(T Item);
-        T RestoreSingle<T>(T Item);
-        IEnumerable<T> RestoreMany<T>(T item);
-    }
-}
